validation, procces error and some fix

+++ controllers/delete_file.controller.js
@@ -1,14 +1,29 @@
-const fileModel = require('../models/file.model')
-const fs = require('fs')
+const fileModel = require("../models/file.model");
+const fs = require("fs");
+const ApiError = require("../helpers/api_error.helper");
 
 const deleteFileController = async (req, res, next) => {
+  try {
+    const filePathFromDB = await fileModel.findOne({
+      where: { user_id: req.user.id, id: req.params.id },
+    });
 
-    const getFilePathFromDB = await fileModel.findOne({where: {user_id: req.user.id, id: req.params.id}})
-    const pathFromDB = getFilePathFromDB.path
-    const removeFileFromStorage = fs.unlinkSync(`./${pathFromDB}`)
+    if(!filePathFromDB) {
+      throw ApiError.BadRequest("Файл не найден");
+    }
 
-    const getFileDB = await fileModel.destroy({where: {user_id: req.user.id, id: req.params.id}})
-    return res.send('success')
+    const pathFromDB = filePathFromDB.path;
+    const fileFromStorage = fs.unlinkSync(`./${pathFromDB}`);
+
+  
+    const deleteFileDB = await fileModel.destroy({
+      where: { user_id: req.user.id, id: req.params.id },
+    });
+    return res.json({message:"Файл успешно удален"});
+  } catch (err) {
+    next(err);
   }
+};
+module.exports = deleteFileController;
 
-module.exports = deleteFileController
+
